### 2021-04-30 #3234 @whedon
[PRE REVIEW]: RENT: A Python Package for Repeated Elastic Net Feature Selection
**Submitting author:** @annajenul (<a href="http://orcid.org/0000-0002-6919-3483">Anna Jenul</a>)
**Repository:** <a href="https://github.com/NMBU-Data-Science/RENT" target ="_blank">https://github.com/NMBU-Data-Science/RENT</a>
**Version:** 0.0.1
