### 2021-05-01 #3234 @annajenul
> **Submitting author:** @annajenul ([Anna Jenul](http://orcid.org/0000-0002-6919-3483))
> **Repository:** https://github.com/NMBU-Data-Science/RENT
> **Version:** 0.0.1
> **Editor:** Pending
> **Reviewer:** Pending
> **Managing EiC:** Kevin M. Moerman
>
> **⚠️ JOSS reduced service mode ⚠️**
>
> Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @annajenul. **Currently, there isn't an JOSS editor assigned** to your paper.
>
> @annajenul if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
>
> **Editor instructions**
>
> The JOSS submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
>
> ```
> @whedon commands
> ```

From your list of potential reviewers, I suggest yxoos, arunmano121 and maximtrp as reviewers for our submission.
