### 2021-05-20 b78d648
Update RENT.py
