### 2021-05-25 #3234 @mikldk
@annajenul Thanks for your submission. (I've actually been to NMBU in Ås a few times -- lovely place!)

I have a few questions. They are more related to the method than the software as such, but they may be relevant for the documentation and paper, so I pose them now before finding reviewers.

* The paper and `README` mention 'unique subsets of the data' -- is this some sort of [bagging](https://en.wikipedia.org/wiki/Bootstrap_aggregating)?
* The paper says 'trains K independent elastic net regularized models on distinct subsets of the train dataset'.
    + I am not sure how these subsets are selected, can you mention this briefly somewhere (maybe both in paper and `README`)?
    + I would doubt that the K subsets are (statistically) independent as you write. Can you support this statement somehow?
* Have you considered also using [feature bagging](https://en.wikipedia.org/wiki/Random_subspace_method)? If using both bagging and feature bagging, the method will be similar to that of [random forests](https://en.wikipedia.org/wiki/Random_forest) except using elastic nets instead of decision trees. I am not aware if something like this have been tried before nor whether there are any advantages/disadvantages. I was just curios that you do not seem to mention 'bagging' nor 'random forests' in the paper (nor the [arXiv](https://arxiv.org/abs/2009.12780) paper).
