### 2021-05-27 #3234 @annajenul
@mikldk  Thank you for your interesting inputs. Glad to hear that you liked your stay at NMBU. The campus is really nice now in the spring / summertime.

Regarding the selection of the K subsets from the training data: we use the scikit-learn train_test_split function K times, every time using a different value for random_state. This gives us K unique subsets of the training data, however without any sample duplicating (compared to what you might get from bootstrapping). We use the term 'unique'  in the sense that none of the K subsets has exactly the same training samples.
https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.train_test_split.html#sklearn.model_selection.train_test_split

In other words, RENT does not use bagging, nor any type of bootstrap. Instead, each training subsample is drawn independently and without replacement from the full training dataset, which means that each sample can appear at most once in a single subset. Nevertheless, the same sample can appear in more than one subset. Therefore, each subset is an iid sample from the training dataset. We will adapt this in the README and the paper to make it more understandable.

Further, we consider to include a bootstrap option as alternative sampling strategy for future work, along with other adaptations, such as with additional classifiers instead of logistic regression. Another option is to apply RENT as a data preprocessing step to tree-based methods which often improved their performance in presence of noisy high cardinal features in the data. In this way we avoided introduction of bias in the tree-based models by removing those noisy high cardinal features (https://explained.ai/rf-importance/index.html).

Although interesting, we have not considered feature bagging, since one ouf our selection criteria (tau_1) is based on counting how often elastic net selects each feature from the full set of features. But this could be an interesting path to follow in future work.
