### 2021-06-01 #3323 @whedon
[REVIEW]: RENT: A Python Package for Repeated Elastic Net Feature Selection
**Submitting author:** @annajenul (<a href="http://orcid.org/0000-0002-6919-3483">Anna Jenul</a>)
**Repository:** <a href="https://github.com/NMBU-Data-Science/RENT" target ="_blank">https://github.com/NMBU-Data-Science/RENT</a>
**Version:** v1.0.0

- [x] #3508
