### 2021-06-02 NMBU-Data-Science/RENT#11 @maximtrp
Restructuring paper
Hello! Thank you for submitting your paper to JOSS! [#3323](https://github.com/openjournals/joss-reviews/issues/3323)

As a reviewer, I suggest restructuring the paper to include separate sections for **Summary** and **Statement of Need**. Please consider these [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

In **Summary** section, describe the functionality and purpose of your package for a *diverse, non-specialist audience*. So, talk about feature selection in general and its current challenges. And then turn to the problem that you are trying to solve.

Your current section **Summary and Statement of Need** can be renamed to **Statement of Need**.
