### 2021-06-03 NMBU-Data-Science/RENT#12 @maximtrp
Community guidelines
Please set up the necessary community guidelines in **Insigths** → **Community** section of your repo (Code of conduct, Contributing, Issue templates). [#3323](https://github.com/openjournals/joss-reviews/issues/3323)
