+++ docs/quickstart.rst
@@ -1,11 +1,11 @@
 Quickstart
 ==========
 
-RENT (Repeated Elastic Net Technique) is a package for feature selection for binary classification problems and regression problems. At its core
-RENT trains an ensemble of unique models using regularized elastic net to select features. Each model in the ensemble is trained with
-a unique and randomly selected subset from the full training data. From these models one can acquire weight distributions for each
+RENT (Repeated Elastic Net Technique) is a feature selection method for binary classification and regression problems. At its core
+RENT trains an ensemble of :math:`K\in\mathbb{N}` generalized linear models using regularized elastic net to select features. Each model :math:`k=1:K` in the ensemble is trained using a randomly, iid sampled subset of rows of the full training data. 
+A single data point can appear at most once in each subset, but may appear in multiple subsets. From these :math:`K` unique models one can acquire weight distributions for each
 feature that contain rich information on the stability of feature selection and from which several adjustable classification criteria may be
-defined.
+defined. 
 
 It is recommended to read the arXiv manuscript `RENT - Repeated Elastic Net Technique for Feature Selection`_, which provides a deeper explanation of the method and is helpful to improve the 
 understanding of RENT and the available analysis methods. 
