### 2021-06-06 NMBU-Data-Science/RENT#11 @annajenul
Hello!
I made the changes you suggested.
Is it ok now?
