### 2021-06-06 NMBU-Data-Science/RENT#11 @maximtrp
Yes, thank you!
