### 2021-06-14 369075f
Create CODE_OF_CONDUCT.md
