+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist =
-   py37
+   py37,py38
 
 [testenv]
 deps =
