### 2021-06-15 bc5f0ba
Create CONTRIBUTING.md
