+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+# Contributing to RENT
+
+Your contribution to RENT is very welcome. For the implementantion of a new feature or bug-fixing, we encourage you to send a Pull Request with a detailed and concise
+description of the invented feature or what the bug was and how you solved it. To improve RENT even more, feel free to send us issues with bugs, you are not sure about. 
+Furthermore, you can also contribute to the improvement of the Read the Docs documentation page. 
+
+When you are finished with your implementantions and bug-fixing, please send a Pull Request to https://github.com/NMBU-Data-Science/RENT.
