+++ tox.ini
@@ -3,6 +3,7 @@ envlist =
    py37,py38
 
 [testenv]
+changedir = tests
 deps =
 	-rrequirements.txt
 commands =
