### 2021-06-15 NMBU-Data-Science/RENT#12 @annajenul
The guidelines are set up now. Does it look ok?
