### 2021-06-18 NMBU-Data-Science/RENT#12 @maximtrp
Yes, looks good!
