### 2021-07-16 #3323 @arunmano121
@mikldk - I have completed the review. In my opinion, the package RENT meets all the reviewer criteria. I was able to install, test and run some sample cases without any issues as described in the documentation. Well done authors - @annajenul and team. Let me know if anything else is needed from my side.
