### 2021-07-18 NMBU-Data-Science/RENT#14 @maximtrp
Documentation: Statement of Need
Please add a section for "Statement of Need" into the documentation [#3323](https://github.com/openjournals/joss-reviews/issues/3323)
