### 2021-07-20 NMBU-Data-Science/RENT#14 @annajenul
A Statement of Need section is now included in the documentation in the section Quickstart.
