### 2021-07-20 NMBU-Data-Science/RENT#14 @maximtrp
Thank you!
